requirements.txt
Flask==3.0.0
Flask-PyMongo==2.3.0
Werkzeug==3.0.1
bcrypt==4.1.1
python-dotenv==1.0.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0
reportlab>=4.0.0
flask-cors>=4.0.0
email-validator>=2.0.0
# Optional ML libraries (install separately if needed)
# tensorflow>=2.15.0
# torch>=2.0.0

